Add files via upload
This is a Data Entry Userform. This form is having mutiple pages. The last page is of the image replacement. This form allows the user to enter the data in the form.When the user clicks on the submit button the the details which are filled in the form gets displayed in the excel sheet. Code is written in VBA and concepts of click events are attached with the various controls in the form